==> 01_wandb_demo.ipynb <==
import wandb
import random

wandb.login()

epochs=10
lr=0.01

run = wandb.init(
    # Set the project where this run will be logged
    project="demo-wandv-project",
    # Track hyperparameters and run metadata
    config={
        "learning_rate": lr,
        "epochs": epochs,
    })

offset = random.random() / 5
print(f"lr: {lr}")

# simulating a training run (accuracy vs. loss)
for epoch in range(2, epochs):
    acc = 1 - 2 ** -epoch - random.random() / epoch - offset
    loss = 2 ** -epoch + random.random() / epoch + offset
    print(f"epoch={epoch}, accuracy={acc}, loss={loss}")
    wandb.log({"accuracy": acc, "loss": loss})

# Creating a log of the "model"
run.log_code()

wandb.finish()
